backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.11.3
pydantic-settings==2.8.
pydantic_core==2.33.1
pydantic[email]==2.11.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
python-multipart==0.0.6
httpx==0.24.1
psycopg2-binary==2.9.6
SQLAlchemy==2.0.40
typing-inspection==0.4.0
redis==5.2.1